refactor(reader): 提取模块到独立文件
